Keep a cache of bound functions for each injector.
This avoids repeatedly re-binding the same function, as well as
guaranteeing that multiple calls can bind the same function and get the
same result.

The cache of bound functions is not inherited by specialized injectors,
to allow for them to override providers and thus achieve a different
result for the same set of dependencies.